comments schema && comments

+++ src/comments/comments.controller.ts
@@ -2,6 +2,7 @@ import {
   BadRequestException,
   Body,
   Controller,
+  Inject,
   Patch,
   Post,
 } from '@nestjs/common';
@@ -25,7 +26,8 @@ import {
   CommentsDeleteOutputDto,
 } from './dtos/comments.delete.dto';
 import {
-  AUTHOR_ID_REQUIRED,
+  BOARD_ID_REQUIRED,
+  NICKNAME_REQUIRED,
   UNIQUE_ID_REQUIRED,
 } from '../_common/constant/errors/400';
 import {
@@ -36,7 +38,9 @@ import {
 @ApiTags('comments')
 @Controller('comments')
 export class CommentsController {
-  constructor(private readonly service: CommentsServiceInterface) {}
+  constructor(
+    @Inject('SERVICE') private readonly service: CommentsServiceInterface,
+  ) {}
 
   @Post('/')
   @ApiConsumes('application/x-www-form-urlencoded')
@@ -45,12 +49,16 @@ export class CommentsController {
     description: '댓글 등록 절차',
   })
   @ApiResponse({ status: 201, description: `${CREATE_SUCCESS}` })
-  @ApiResponse({ status: 400, description: `${AUTHOR_ID_REQUIRED}` })
+  @ApiResponse({
+    status: 400,
+    description: `${BOARD_ID_REQUIRED}, ${NICKNAME_REQUIRED}`,
+  })
   @ApiResponse({ status: 500, description: `${INTERNAL_SERVER_ERROR}` })
   private async register(
     @Body() dto: CommentsRegisterInputDto,
   ): Promise<CommentsRegisterOutputDto> {
-    if (!dto?.authorId) throw new BadRequestException(AUTHOR_ID_REQUIRED);
+    if (!dto?.boardId) throw new BadRequestException(BOARD_ID_REQUIRED);
+    if (!dto?.nickname) throw new BadRequestException(NICKNAME_REQUIRED);
 
     return await this.service.register(dto);
   }
@@ -62,13 +70,17 @@ export class CommentsController {
     description: '댓글 삭제 절차',
   })
   @ApiResponse({ status: 200, description: `${TWO_HUNDRED_OK}` })
-  @ApiResponse({ status: 400, description: `${UNIQUE_ID_REQUIRED}` })
+  @ApiResponse({
+    status: 400,
+    description: `${UNIQUE_ID_REQUIRED}, ${BOARD_ID_REQUIRED}`,
+  })
   @ApiResponse({ status: 404, description: `${NOTFOUND_COMMENT}` })
   @ApiResponse({ status: 500, description: `${INTERNAL_SERVER_ERROR}` })
   private async delete(
     @Body() dto: CommentsDeleteInputDto,
   ): Promise<CommentsDeleteOutputDto> {
     if (!dto?.id) throw new BadRequestException(UNIQUE_ID_REQUIRED);
+    if (!dto?.boardId) throw new BadRequestException(BOARD_ID_REQUIRED);
 
     return await this.service.delete(dto);
   }
@@ -82,7 +94,7 @@ export class CommentsController {
   @ApiResponse({ status: 200, description: `${TWO_HUNDRED_OK}` })
   @ApiResponse({
     status: 400,
-    description: `${UNIQUE_ID_REQUIRED}, ${AUTHOR_ID_REQUIRED}`,
+    description: `${UNIQUE_ID_REQUIRED}, ${BOARD_ID_REQUIRED}`,
   })
   @ApiResponse({ status: 404, description: `${NOTFOUND_COMMENT}` })
   @ApiResponse({ status: 500, description: `${INTERNAL_SERVER_ERROR}` })
@@ -90,7 +102,7 @@ export class CommentsController {
     @Body() dto: CommentsUpdateInputDto,
   ): Promise<CommentsUpdateOutputDto> {
     if (!dto?.id) throw new BadRequestException(UNIQUE_ID_REQUIRED);
-    if (!dto?.authorId) throw new BadRequestException(AUTHOR_ID_REQUIRED);
+    if (!dto?.boardId) throw new BadRequestException(BOARD_ID_REQUIRED);
 
     return await this.service.update(dto);
   }
